app/requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dbus-next==0.2.3
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-socks==0.10.0
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
mss==10.0.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycairo==1.27.0
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pygobject==3.52.3
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-socks==2.7.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
triton==3.2.0
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
